data_process/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
pandas==2.2.3
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1